+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# v4.36.9 
+# v4.37.1 
+🚀Private Release - 4.37.1 🚀
+MercadoPagoSDKV4 - Private Version
+- Remove groups and unused assets
+
+# v4.37.0
+🚀Private Release - 4.37.0 🚀
+MercadoPagoSDKV4 - Private Version
+- Add hybrid card support
+
+# v4.36.9
 🚀Private Release - 4.36.9 🚀
 MercadoPagoSDKV4 - Private Version
 - Added location header to congrats